Re-organized tests so unittest can find them by default.
