+++ runtests.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+python -m unittest discover --start-directory search --top-level-directory .
