+++ search/tests/__init__.py
@@ -1,2 +1,2 @@
-from fields import *
-from indexers import *
+from test_fields import *
+from test_indexers import *
